// src/api/models/getStoreTicketListResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Allta API
 * Allta API 서버 문서입니다.
 * OpenAPI spec version: 1.0
 */
import type { StoreTicketListItem } from './storeTicketListItem';
import type { PaginationMetaEntity } from './paginationMetaEntity';

export interface GetStoreTicketListResponse {
  /** 결과 여부 */
  ok: boolean;
  /** 일회권 목록 */
  data: StoreTicketListItem[];
  /** 메타 정보 */
  meta: PaginationMetaEntity;
}